feat: Add code for inviting more guests to dinner
This commit adds code to invite more guests to the dinner. It creates a new file called moreGuests.py and includes a list of existing guests. The code informs about a bigger table and adds new guests to the list using different methods such as inserting at the beginning and middle, as well as appending to the end. Finally, it sends new invitations to all the guests.