feat: Add code for plotting and saving a figure of square numbers
This commit adds code to plot a graph of square numbers using matplotlib and save the figure as an image file. The code creates a line plot of the square numbers and sets the chart title, axes labels, and tick label sizes. The SaveFig class is used to save the figure as "mpl_squares.png".

Ref: #issue-number