feat: Add code for generating and saving figures of a D6 die roll and a random walk
This commit adds code to generate and save figures of a D6 die roll and a random walk. The `roll_die` function rolls a D6 die a specified number of times and plots the results. The `visualize_random_walk` function creates and visualizes a random walk using Plotly. The generated figures are saved as `d6_plot.png` and `random_walk_plotly.html`, respectively.